Rough cut of functionality finished
It actually does something now! Data gets copied and appears on the
clips menu, and clicking on one will result in that data being pasted.
There are some weird timing issue that I'm still trying to work out. I'm
pretty sure they are a combination of windows not finishing closing the
clipboard, and trying to send the Ctrl+v too quickly.